+++ src/waymark/__init__.py
@@ -1,4 +1,3 @@
 """Waymark package."""
 
-__version__ = "0.1.0"
-
+__version__ = "0.2.0.dev0"